Extract prepared texture uploader
